+++ changelog.md
@@ -1,5 +1,9 @@
 # eslint-config-env changelog
 
+## Next
+
+- Fix a crash when a package is missing `peerDependencies`, `dependencies` or `devDependencies` fields.
+
 ## 0.1.0
 
 - Initial release.